Move escaping routines into a shared maud_htmlescape crate
