+++ maud_htmlescape/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+
+name = "maud_htmlescape"
+# When releasing a new version, please update html_root_url in lib.rs
+version = "0.17.0"
+authors = ["Chris Wong <lambda.fairy@gmail.com>"]
+
+license = "MIT/Apache-2.0"
+documentation = "https://docs.rs/maud_htmlescape/"
+homepage = "https://maud.lambda.xyz/"
+repository = "https://github.com/lfairy/maud"
+description = "Internal support code used by Maud."
+
+[lib]
+path = "lib.rs"
